+++ main/trans/old_model/shop_user_role.go
@@ -8,12 +8,12 @@ import (
 )
 
 type ShopUserRole struct {
-	ID         uint `gorm:"primary_key;AUTO_INCREMENT;comment:主键id"`
-	ShopUserID uint `gorm:"not null;default:0;comment:超管用户id"`
-	RoleID     uint `gorm:"not null;default:0;comment:角色id"`
-	AppID      uint `gorm:"not null;default:0;comment:小程序id"`
-	CreateTime int  `gorm:"not null;default:0;comment:创建时间"`
-	UpdateTime int  `gorm:"default:0;comment:更新时间"`
+	ID         uint  `gorm:"primary_key;AUTO_INCREMENT;comment:主键id"`
+	ShopUserID uint  `gorm:"not null;default:0;comment:超管用户id"`
+	RoleID     uint  `gorm:"not null;default:0;comment:角色id"`
+	AppID      uint  `gorm:"not null;default:0;comment:小程序id"`
+	CreateTime int64 `gorm:"not null;default:0;comment:创建时间"`
+	UpdateTime int64 `gorm:"default:0;comment:更新时间"`
 }
 type ShopUserRoleRepository interface {
 	GetShopUserRoleList() ([]*ShopUserRole, error)
